Move out types from db package
