+++ packages/types/src/index.ts
@@ -1,2 +1,3 @@
 export * from './discord/event';
 export * from './discord/command';
+export * from './db/mongodb/types';
